color

+++ colorize.py
@@ -54,7 +54,7 @@ if __name__=='__main__':
     parser.add_argument('--data_path', type=str, default = '/data/landscapes', help='data path')
     parser.add_argument('--batch_size', type=int, default = 1, help='batch size')
     parser.add_argument('--epochs', type=int, default = 1, help='epochs')
-    parser.add_argument('--lr', type=int, default = 1e-3, help='learning rate')
+    parser.add_argument('--lr', type=float, default = 1e-3, help='learning rate')
 
     exp_name = 'colorize'
     args = parser.parse_args()
